saperate code
